Default Text Style、Safe Areaの追加、細かなデザインの修正
コードの整理もしました
