--- lib/youtube/components/youtube_videos_section_title.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-
-class YoutubeVideosSectionTitle extends StatelessWidget {
-  const YoutubeVideosSectionTitle(this.title, {Key? key}) : super(key: key);
-
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(
-        vertical: 20,
-        horizontal: 20,
-      ),
-      child: Text(
-        title,
-        style: const TextStyle(
-          color: Colors.white,
-          fontSize: 18,
-        ),
-      ),
-    );
-  }
-}
